Minor updates: 1) S3Service methods changed to throw Exception instead of throwing list of exact minio exceptions 2) Custom exception handler and ErrorResponse class removed.

+++ fileStorageService/src/main/java/ylab/bies/fileStorageService/service/impl/FileServiceImpl.java
@@ -83,6 +83,12 @@ public class FileServiceImpl implements FileService {
   public void removeFile(UUID fileId) {
     FileModel fileModelToRemove = fileRepository.findById(fileId)
             .orElseThrow(RequestedFileNotFoundException::new);
+    long ideaId = fileModelToRemove.getIdeaId();
+
+    String dummyToken = "test";
+    if (dummyToken == null || !ideaServiceClient.validateIdeaOwner(ideaId, dummyToken)) {
+      throw new IdeaOwnershipException(ideaId);
+    }
 
     fileRepository.delete(fileModelToRemove);
